harmonize blitstr2 and graphics-server to an i16 pixel representation
previously blitstr2 is using usize for x/y coords, which means that
for layout purposes points could not go to the top or left boundary
without causing arithmetic overflow. Of course, you want to render
everything on-screen, but, as we have multi-line text with overflow
possibility, it's reasonable to be able to use negative coordinates
to locate text boxes that flow into the visible region.

This of course means that occassionally you can come out with
things like "negative widths", but even then, there can be reasonable
uses for this as an intermediate in calculations that ultimately
arrive at a positive result.

+++ services/graphics-server/src/api/blitstr2.rs
@@ -42,9 +42,9 @@ pub(crate) struct TypesetWord {
     /// top left origin point for rendering of the glyphs
     pub origin: Pt,
     /// width of the word
-    pub width: usize,
+    pub width: i16,
     /// overall height for the word
-    pub height: usize,
+    pub height: i16,
     /// set if this `word` is not drawable, e.g. a newline placeholder.
     /// *however* the Vec<GlyphSprite> should still be checked for an insertion point, so that
     /// successive newlines properly get their insertion point drawn
